Uploaded another version with all coins same csv
It is important to see that there are more data points in the 4_coins.csv file than in the individual coin files. The reason is that we haven't set a limit on the end date, we can easily change it. So far the code retrieves values from 20-08-2020 to the present time.